kcov: Add kcov and sancov support
Add kcov, a kernel coverage device that leverages clang/gcc sancov to
provide coverage data to userspace. The kcov interface follows
Linux/FreeBSD/NetBSD/OpenBSD's semantics.

trace-cmp mode is still not implemented.

+++ kernel/lib/libk/include/bits/ioctl.h
@@ -1,4 +1,4 @@
-#define _IOC(a,b,c,d) ( ((a)<<30) | ((b)<<8) | (c) | ((d)<<16) )
+#define _IOC(a,b,c,d) ((int)( ((a)<<30) | ((b)<<8) | (c) | ((d)<<16) ))
 #define _IOC_NONE  0U
 #define _IOC_WRITE 1U
 #define _IOC_READ  2U
